src/requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pypdfium2==5.1.0
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
